+++ src/backend/chat/clients/pydantic_ai.py
@@ -72,6 +72,7 @@ from chat.clients.pydantic_ui_message_converter import (
     ui_message_to_user_content,
 )
 from chat.mcp_servers import get_mcp_servers
+from chat.tools.document_generic_search_rag import add_document_rag_search_tool_from_setting
 from chat.tools.document_search_rag import add_document_rag_search_tool
 from chat.tools.document_summarize import document_summarize
 from chat.vercel_ai_sdk.core import events_v4, events_v5
@@ -140,6 +141,7 @@ class AIAgentService:  # pylint: disable=too-many-instance-attributes
             instrument=self._store_analytics,
             deps_type=ContextDeps,
         )
+        add_document_rag_search_tool_from_setting(self.conversation_agent, self.user)
 
     @property
     def _stop_cache_key(self):